+++ frontend/src/routes/post/[post_id]/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { enhance } from '$app/forms';
 	import CommentTree from '$lib/CommentTree.svelte';
-	import EditBy from '$lib/EditBy.svelte';
+	import EditedBy from '$lib/EditedBy.svelte';
 	import LangSwitch from '$lib/LangSwitch.svelte';
 	import Section from '$lib/Section.svelte';
 	import TimeAgo from '$lib/TimeAgo.svelte';
@@ -195,7 +195,7 @@
 					>
 					<a href="#p{data.post_id}"><TimeAgo date={page_object.modified} /></a>
 					{#if data.post.edited_at && data.post.edited_by}
-						<EditBy
+						<EditedBy
 							date={data.post.edited_at}
 							user={editedByOther ? data.post.edited_by : null}
 						/>
